Zip component is divided to interface + Impl
