+++ apl-exec/src/main/java/com/apollocurrency/aplwallet/apl/exec/webui/WebUiExtractor.java
@@ -10,6 +10,7 @@ import java.nio.file.Paths;
 import java.util.concurrent.Callable;
 
 import com.apollocurrency.aplwallet.apl.util.Zip;
+import com.apollocurrency.aplwallet.apl.util.ZipImpl;
 import com.apollocurrency.aplwallet.apl.util.env.dirprovider.DirProvider;
 
 /**
@@ -111,7 +112,7 @@ public class WebUiExtractor implements Callable<Boolean>{
     
     public boolean install() throws FileNotFoundException, IOException {
         boolean res=true;
-        Zip zip = new Zip();
+        Zip zip = new ZipImpl();
         if(!checkInstalled()){
             File dest = findDest();
             if(dest.exists()){
